pyproject.toml
[tool.poetry]
name = "back-end"
version = "0.1.0"
description = ""
authors = ["salehmmasri <salehalmasri494@yahoo.com>"]

[tool.poetry.dependencies]
python = "^3.8"
djangorestframework = "^3.12.2"
django-cors-headers = "^3.6.0"
django = "^3.1.5"
django-phonenumber-field = "^5.0.0"
phonenumbers = "^8.12.17"
psycopg2-binary = "^2.8.6"
django-environ = "^0.4.5"
gunicorn = "^20.0.4"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
django = "^3.1.5"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
